requirements.txt
Django==1.11.6
django-cors-headers==2.1.0
django-extensions==1.9.6
django-model-utils==3.0.0
djangorestframework==3.7.1
-e git+https://github.com/night-crawler/drf-metadata.git@1b1a78725169466baa70085ae19414ea1873bcaf#egg=drf_metadata
easy-thumbnails==2.4.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.2
six==1.11.0
